safari-600.7-branch/LayoutTests/platform/mac/editing/input/edit-dictated-text-with-alternative-expected.txt

Editing text that has DictationAlternatives marker whould cause the marker to be removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

this is a tex
PASS markedText is ""

TEST COMPLETE
